+++ src/components/ShortDropdown/ShortDropdown.tsx
@@ -1,6 +1,6 @@
-import ShortDropdownList from "./shortDropdownList";
-import ShortDropdownMenu from "./shortDropdownMenu";
-import ShortDropdownTrigger from "./shortDropdownTrigger";
+import ShortDropdownList from "@/src/components/ShortDropdown/ShortDropdownList";
+import ShortDropdownMenu from "@/src/components/ShortDropdown/ShortDropdownMenu";
+import ShortDropdownTrigger from "@/src/components/ShortDropdown/ShortDropdownTrigger";
 import useDetectClose from "@/src/hooks/useDetectClose";
 import clsx from "clsx";
 import { ReactNode } from "react";
